+++ gui_main.py
@@ -1,5 +1,5 @@
 """Helper script used by Buildozer when generating an Android APK."""
-from stattrackerpro.gui.app import main
+from saavygambler.gui.app import main
 
 
 if __name__ == "__main__":  # pragma: no cover - manual execution
