+++ saavygambler/__init__.py
@@ -1,4 +1,4 @@
-"""StatTrackerPro package initialization."""
+"""SaavyGambler package initialization."""
 from __future__ import annotations
 
 from importlib import import_module
