+++ saavygambler/gui/__init__.py
@@ -1,0 +1,3 @@
+"""Graphical interface package for SaavyGambler."""
+
+__all__ = []
