+++ saavygambler/providers/base.py
