--- stattrackerpro/gui/__init__.py
@@ -1,3 +1,0 @@
-"""Graphical interface package for StatTrackerPro."""
-
-__all__ = []
